[Gardening] Random fixes to the type system
Mostly, adding comments for things that are unlikely to change.

Also:
* Move Instantiation to a separate file. We generally keep helper
structs together with classes that use them, but this one is used by
generic methods too and doesn't feel like something intrinsic to
TypeDesc.
* Delete useless overrides on ArrayMethod. Base implementation is
identical.
* Add a comment about weird overrides of GetTypeDefinition. Their
purpose puzzles me - we should either document it, or delete them. I
don't think anything relies on the sematic.
