+++ src/Common/src/TypeSystem/Common/DefType.cs
@@ -2,13 +2,11 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-using System;
-
 namespace Internal.TypeSystem
 {
     /// <summary>
     /// Type that is logically equivalent to a type which is defined by a TypeDef
-    /// record in an ECMA 335 metadata stream.
+    /// record in an ECMA 335 metadata stream - a class, an interface, or a value type.
     /// </summary>
     public abstract partial class DefType : TypeDesc
     {
