+++ src/Common/src/TypeSystem/Common/PointerType.cs
@@ -2,10 +2,11 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-using System;
-
 namespace Internal.TypeSystem
 {
+    /// <summary>
+    /// Represents an unmanaged pointer type.
+    /// </summary>
     public sealed partial class PointerType : ParameterizedType
     {
         internal PointerType(TypeDesc parameterType)
